fix: pre-existing import sort and scipy-stubs typing nit
ruff check . flagged I001 in latent_extrapolation_preprocessing_ab.py
(from 4e1b9f1); installing scipy-stubs then surfaced a max(0.0, scalar)
union-type error in energy_distance_multivariate. Both fixes are
behavior-preserving; mypy src and ruff check . are now clean.

+++ src/reuse_gate/metrics/distribution.py
@@ -28,7 +28,7 @@ def energy_distance_multivariate(x: npt.NDArray[Any], y: npt.NDArray[Any]) -> fl
     b = np.mean(d_xx)  # within-x
     c = np.mean(d_yy)  # within-y
 
-    return float(max(0.0, 2.0 * a - b - c))
+    return float(max(0.0, float(2.0 * a - b - c)))
 
 
 def median_pairwise_distance(
